Adjust 02-flex-header exercise for completion
Adjusted padding and margin values to match the solution. Removed all
commented out and unnecessary properties (flex). Added gap to ul and
removed margin from a. Original solution was very close, but looked
incorrect because of the limitations in the viewing window size in the
VSCode preview (was below the minimum size and crushed elements).